requirements/base.txt
fastapi==0.109.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3
python-dotenv==1.0.0
fastapi_redis==0.1.11
openpyxl==3.1.2
celery==5.3.6
uvicorn==0.26.0
loguru==0.7.2
